requirements.txt
streamlit
google-generativeai
gspread
google-auth